# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "znskr api",
    "description": "znskr paas api for managing apps and deployments",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/apps": {
      "get": {
        "tags": [
          "apps"
        ],
        "summary": "list all apps for the authenticated user",
        "operationId": "list_apps",
        "responses": {
          "200": {
            "description": "list of apps",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "apps"
        ],
        "summary": "create a new app",
        "operationId": "create_app",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "app created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "domain conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/apps/{id}": {
      "get": {
        "tags": [
          "apps"
        ],
        "summary": "get a single app by id",
        "operationId": "get_app",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "app id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "app details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "apps"
        ],
        "summary": "update an app",
        "operationId": "update_app",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "app id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "app updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "domain conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "apps"
        ],
        "summary": "delete an app",
        "operationId": "delete_app",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "app id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "app deleted"
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/apps/{id}/git": {
      "get": {
        "tags": [
          "git"
        ],
        "summary": "get git push info for an app",
        "operationId": "get_git_info",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "app id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "git info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitInfoResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "git"
        ],
        "summary": "enable git push for an app",
        "operationId": "enable_git",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "app id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "git enabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitEnableResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "already enabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/github/app": {
      "get": {
        "tags": [
          "github-app"
        ],
        "summary": "get github app status and installations",
        "operationId": "get_github_app",
        "responses": {
          "200": {
            "description": "github app status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubAppStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "github-app"
        ],
        "summary": "delete github app configuration",
        "operationId": "delete_github_app",
        "responses": {
          "200": {
            "description": "github app deleted"
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/github/app/callback": {
      "get": {
        "tags": [
          "github-app"
        ],
        "summary": "handles the github app manifest callback",
        "operationId": "github_app_callback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "manifest conversion code",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "redirect to settings"
          },
          "400": {
            "description": "conversion failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/github/app/install/callback": {
      "get": {
        "tags": [
          "github-app"
        ],
        "summary": "handles the installation callback",
        "operationId": "github_install_callback",
        "parameters": [
          {
            "name": "installation_id",
            "in": "query",
            "description": "installation id",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "setup_action",
            "in": "query",
            "description": "setup action",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "redirect to settings"
          }
        }
      }
    },
    "/api/github/app/manifest": {
      "get": {
        "tags": [
          "github-app"
        ],
        "summary": "generates the github app manifest and returns the creation url",
        "operationId": "get_app_manifest",
        "responses": {
          "200": {
            "description": "app manifest payload",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/github/app/repos": {
      "get": {
        "tags": [
          "github-app"
        ],
        "summary": "get repos from github app installations",
        "operationId": "get_app_repos",
        "responses": {
          "200": {
            "description": "list of repositories",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppReposResponse"
                }
              }
            }
          },
          "400": {
            "description": "no github app configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppReposResponse": {
        "type": "object",
        "description": "repos response",
        "required": [
          "repos"
        ],
        "properties": {
          "repos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RepoInfo"
            }
          }
        }
      },
      "AppResponse": {
        "type": "object",
        "description": "app response",
        "required": [
          "id",
          "name",
          "github_url",
          "branch",
          "domains",
          "port",
          "env_vars",
          "services",
          "rollout_strategy",
          "created_at"
        ],
        "properties": {
          "branch": {
            "type": "string",
            "description": "branch being deployed"
          },
          "created_at": {
            "type": "string",
            "description": "creation timestamp"
          },
          "domain": {
            "type": [
              "string",
              "null"
            ],
            "description": "custom domain"
          },
          "domains": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "custom domains"
          },
          "env_vars": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnvVarResponse"
            },
            "description": "environment variables"
          },
          "github_url": {
            "type": "string",
            "description": "github repository url"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "unique app id"
          },
          "name": {
            "type": "string",
            "description": "app name"
          },
          "port": {
            "type": "integer",
            "format": "int32",
            "description": "app port (deprecated)",
            "minimum": 0
          },
          "rollout_strategy": {
            "type": "string",
            "description": "rollout strategy"
          },
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceResponse"
            },
            "description": "container services"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "error response",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "error message"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}